+++ public/js/app.js
@@ -5,14 +5,14 @@
 // ═══════════════════════════════════════════════════════════
 
 import SocketMethods   from './modules/app-socket.js?v=4.5.5';
-import UIBindMethods   from './modules/app-ui.js?v=4.5.4';
+import UIBindMethods   from './modules/app-ui.js?v=4.6.1';
 import MediaMethods    from './modules/app-media.js?v=4.5.4';
 import ContextMethods  from './modules/app-context.js?v=3.44.6';
-import ChannelMethods  from './modules/app-channels.js?v=4.5.4';
+import ChannelMethods  from './modules/app-channels.js?v=4.6.1';
 import MessageMethods  from './modules/app-messages.js?v=4.5.1';
 import UserMethods     from './modules/app-users.js?v=4.5.1';
 import VoiceMethods    from './modules/app-voice.js?v=4.5.1';
-import UtilityMethods  from './modules/app-utilities.js?v=4.5.3';
+import UtilityMethods  from './modules/app-utilities.js?v=4.6.1';
 import AdminMethods    from './modules/app-admin.js?v=4.5.1';
 import PlatformMethods from './modules/app-platform.js?v=4.5.1';
 import SearchMethods   from './modules/app-search.js?v=3.49.0';